+++ src/Farmer/Arm/Web.fs
@@ -10,6 +10,7 @@ let sites = ResourceType ("Microsoft.Web/sites", "2016-08-01")
 let config = ResourceType ("Microsoft.Web/sites/config", "2016-08-01")
 let sourceControls = ResourceType ("Microsoft.Web/sites/sourcecontrols", "2019-08-01")
 let staticSites = ResourceType("Microsoft.Web/staticSites", "2019-12-01-preview")
+let siteExtensions = ResourceType("Microsoft.Web/sites/siteextensions", "2020-06-01")
 
 type ServerFarm =
     { Name : ResourceName
@@ -30,7 +31,7 @@ type ServerFarm =
     member this.Kind =
         match this.OperatingSystem with
         | Linux -> Some "linux"
-        | _ -> None
+        | Windows -> None
     member this.Tier =
         match this.Sku with
         | Free -> "Free"
@@ -245,3 +246,13 @@ type StaticSite =
             this.RepositoryToken
         ]
 
+[<AutoOpen>]
+module SiteExtensions =
+    type SiteExtension =
+        { Name : ResourceName
+          SiteName : ResourceName
+          Location : Location }
+        interface IArmResource with
+            member this.ResourceId = siteExtensions.resourceId(this.SiteName/this.Name)
+            member this.JsonModel =
+                siteExtensions.Create(this.SiteName/this.Name, this.Location, [ sites.resourceId this.SiteName ]) :> _